+++ src/main/java/com/goldenfield192/irpatches/document/core/DefaultPageBuilder.java
@@ -1,16 +1,16 @@
-package com.goldenfield192.irpatches.document.markdown;
+package com.goldenfield192.irpatches.document.core;
 
 import cam72cam.mod.resource.Identifier;
 import com.goldenfield192.irpatches.document.manual.GaugeProvider;
 import com.goldenfield192.irpatches.document.manual.StockListProvider;
 import com.goldenfield192.irpatches.document.manual.TrackProvider;
-import com.goldenfield192.irpatches.document.markdown.element.*;
+import com.goldenfield192.irpatches.document.core.element.*;
 
 import java.util.*;
 import java.util.function.BiFunction;
 
-import static com.goldenfield192.irpatches.document.markdown.element.MarkdownStyledText.MARKER_PARSE_PRIORITY;
-import static com.goldenfield192.irpatches.document.markdown.element.MarkdownStyledText.MARKER_STYLES;
+import static com.goldenfield192.irpatches.document.core.element.MarkdownStyledText.MARKER_PARSE_PRIORITY;
+import static com.goldenfield192.irpatches.document.core.element.MarkdownStyledText.MARKER_STYLES;
 
 /**
  * Load Markdown from existing file
@@ -39,9 +39,9 @@ public class DefaultPageBuilder implements IPageBuilder {
     }
 
     /**
-     * Internal method to parse a raw String markdown line into MarkdownStyledText and MarkdownUrl
+     * Internal method to parse a raw String core line into MarkdownStyledText and MarkdownUrl
      *
-     * @param input Raw markdown line
+     * @param input Raw core line
      * @return Parsed AbstractMarkdownElement line
      */
     public static List<AbstractMarkdownElement> parse(String input) {
@@ -196,7 +196,7 @@ public class DefaultPageBuilder implements IPageBuilder {
     /**
      * Builds a MarkdownDocument from the given resource identifier
      *
-     * @param id The identifier containing markdown content
+     * @param id The identifier containing core content
      * @return Built MarkdownDocument instance
      */
     @Override