+++ src/main/java/com/goldenfield192/irpatches/document/core/element/AbstractMarkdownElement.java
@@ -1,4 +1,4 @@
-package com.goldenfield192.irpatches.document.markdown.element;
+package com.goldenfield192.irpatches.document.core.element;
 
 import cam72cam.mod.render.opengl.RenderState;