+++ src/main/java/com/goldenfield192/irpatches/document/core/element/MarkdownSplitLine.java
@@ -1,11 +1,11 @@
-package com.goldenfield192.irpatches.document.markdown.element;
+package com.goldenfield192.irpatches.document.core.element;
 
 import cam72cam.mod.gui.helpers.GUIHelpers;
 import cam72cam.mod.math.Vec3d;
 import cam72cam.mod.render.opengl.RenderState;
 import com.goldenfield192.irpatches.IRPConfig;
 
-import static com.goldenfield192.irpatches.document.markdown.Colors.SPLIT_LINE_COLOR;
+import static com.goldenfield192.irpatches.document.core.Colors.SPLIT_LINE_COLOR;
 
 /**
  * Element class representing a split line