+++ src/main/java/com/goldenfield192/irpatches/document/core/element/MarkdownTitle.java
@@ -1,9 +1,9 @@
-package com.goldenfield192.irpatches.document.markdown.element;
+package com.goldenfield192.irpatches.document.core.element;
 
 import cam72cam.mod.render.opengl.RenderState;
 import com.goldenfield192.irpatches.gui.IRPGUIHelper;
 
-import static com.goldenfield192.irpatches.document.markdown.Colors.DEFAULT_TEXT_COLOR;
+import static com.goldenfield192.irpatches.document.core.Colors.DEFAULT_TEXT_COLOR;
 
 /**
  * Element class representing a title