+++ src/main/java/com/goldenfield192/irpatches/document/manual/element/MDStockModelRenderer.java
@@ -6,7 +6,7 @@ import cam72cam.mod.math.Vec3d;
 import cam72cam.mod.render.obj.OBJRender;
 import cam72cam.mod.render.opengl.RenderState;
 import com.goldenfield192.irpatches.IRPConfig;
-import com.goldenfield192.irpatches.document.markdown.element.AbstractMarkdownElement;
+import com.goldenfield192.irpatches.document.core.element.AbstractMarkdownElement;
 
 import java.util.ArrayList;
 import java.util.Arrays;