+++ src/main/java/com/goldenfield192/irpatches/mixins/minecraft/MixinEntityRenderer.java
@@ -6,6 +6,7 @@ import net.minecraft.client.renderer.EntityRenderer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.Constant;
 import org.spongepowered.asm.mixin.injection.ModifyConstant;
+import org.spongepowered.asm.mixin.injection.Surrogate;
 
 @Mixin(EntityRenderer.class)
 public class MixinEntityRenderer {